requirement.txt
astroid==2.2.5
colorama==0.4.1
Django==2.2.4
django-tinymce4-lite==1.7.5
gunicorn==19.9.0
isort==4.3.21
jsmin==2.2.2
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.1.0
psycopg2-binary==2.8.4
pylint==2.3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
